fix(HandlerVitalsScheduler): HandlerVitalsScheduler now fully encapsulates its `HandlerThread` and `Handler` instead of these being constructed in `VitalsDataSource`
