+++ embrace-android-instrumentation-vitals/src/main/kotlin/io/embrace/android/embracesdk/internal/vitals/VitalsDataSource.kt
@@ -3,8 +3,6 @@ package io.embrace.android.embracesdk.internal.vitals
 import android.content.Context
 import android.hardware.display.DisplayManager
 import android.os.Build
-import android.os.Handler
-import android.os.HandlerThread
 import android.view.Display
 import androidx.annotation.RequiresApi
 import io.embrace.android.embracesdk.internal.arch.InstrumentationArgs
@@ -44,14 +42,13 @@ internal class VitalsDataSource(
         override fun onDisplayRemoved(displayId: Int) {}
     }
 
-    private var handlerThread: HandlerThread? = null
-    private var scheduler: VitalsScheduler? = null
+    private var scheduler: HandlerVitalsScheduler? = null
     private var activityListener: VitalsActivityListener? = null
 
     override fun onDataCaptureEnabled() {
-        val thread = HandlerThread(HANDLER_THREAD_NAME).apply { start() }
-        handlerThread = thread
-        val handler = Handler(thread.looper)
+        val vitalsScheduler = HandlerVitalsScheduler().apply { start() }
+        scheduler = vitalsScheduler
+        val handler = vitalsScheduler.handler
 
         // Below API 31 the budget comes from [refreshRate]; track refresh-rate changes to keep it live.
         if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) {
@@ -59,8 +56,6 @@ internal class VitalsDataSource(
                 ?.registerDisplayListener(displayListener, handler)
         }
 
-        val vitalsScheduler = HandlerVitalsScheduler(handler)
-        scheduler = vitalsScheduler
         val tracker = FocalMomentTracker(
             scheduler = vitalsScheduler,
             reporter = SmoothnessReporter(emit = ::emitSmoothnessResult),
@@ -84,11 +79,9 @@ internal class VitalsDataSource(
             args.systemService<DisplayManager>(Context.DISPLAY_SERVICE)
                 ?.unregisterDisplayListener(displayListener)
         }
-        // Cancels any pending settle for an open focal moment.
-        scheduler?.cancelSettle()
+        // Cancels any pending settle for an open focal moment and stops the vitals thread.
+        scheduler?.stop()
         scheduler = null
-        handlerThread?.quitSafely()
-        handlerThread = null
     }
 
     private fun emitSmoothnessResult(result: SmoothnessResult) {
@@ -120,8 +113,6 @@ internal class VitalsDataSource(
     }
 
     private companion object {
-        const val HANDLER_THREAD_NAME = "emb-vitals"
-
         /**
          * Our "normalized" refresh rate is locked at 60fps, we adjust all frame rates to this so 1 frame on a 30fps screen counts as
          * 2 frames and 1 frame on a 120fps display rate is considered 0.5 frames.
